pyproject.toml
[project]
name = "duq-ds3-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "Sara Kelemen"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.1,<3.0.0)",
    "scipy (>=1.15.1,<2.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "xgboost (>=2.1.3,<3.0.0)",
    "instaloader (>=4.14.1,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "fuzzywuzzy (>=0.18.0,<0.19.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "levenshtein (>=0.27.1,<0.28.0)",
    "jarowinkler (>=2.0.1,<3.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "fasttext-wheel (>=0.9.2,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
